+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,11 +1,11 @@
-package kitchenpos.product.dao;
+package kitchenpos.product.repository;
 
 import kitchenpos.product.model.Product;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductDao {
+public interface ProductRepository {
     Product save(Product entity);
 
     Optional<Product> findById(Long id);
